Stop recording new VCR episodes for each test run
Due to the order and payment having new identifiers on each run, we
would never match the body of the request. Since we had VCR configured
to record new episodes by default, it would just keep appending to the
end of the file.

This ensures the body will match by always using the same identifiers
and also removes the global config option to record new episodes.