Default to US country code
There's an issue with Apple Pay in which they won't send us a country if
a customer decides to use an existing address. As a semi-elegant
solution, we assume that the customer is from the USA.

This allows us to continue as if nothing went wrong for all of those
customers. And any not from the USA will see the same error they saw
before. (Due to whatever state they specified not being valid for the
USA.)