Pass configuration object to initialize
With submitting transactions through the API we should not rely on the
transaction form being present on the page. That means we need to pass
in the email somewhere if we don't want to prompt for it each time. We
also need to provide the payment method ID unless it's set globally.

The list of parameters we needed in initialize was getting so long, I
figured it made sense to provide a single configuration object with all
these values. This way for things like the email, we can optionally pass
it in.