Refactor frontend PayPal JS
This tries to simplify what was happening before. I created an
instantiable PayPal object that encapsulates most of the logic for
creating the button. Instead of hardcoding the options passed to the
paypal tokenize call in initializePaypal and initializePaypalCredit, I
just use one function that allows the user to pass in the configuration
options.