Make the gem a mountable engine to isolate routes
Currently we are appending all our routes to the host application's
routes, which can cause conflicts with existing ones.

This change makes it so the gem's engine should be mounted in the host
application's config/routes.rb and keeps the engine's routes inside of
their own namespace. The host application can still use the gem's
routes, but it must refer to them explicitly using the routing proxy.
ie: `solidus_paypal_braintree.store_configuration_path` vs
`store_configuration_path`.

One issue encountered is that Solidus doesn't always use the `spree.`
routing proxy explicitly. If a page is rendered from a controller inside
the gem, any url helpers that don't explicitly use a routing proxy will
be looked up in the gem's routes causing errors. The RoutesHelper module
was added to deal with that. When included in a controller, the
method_missing will retry the url helper using the spree routing proxy
and return that result if it exists, otherwise it will raise the error.