Fix error when saving braintree payment method
If you try to update the Braintree payment method in the admin it will
throw an error due to 2 of the preference types being hashes. Solidus
automatically creates a text field for these on the page, which it
populates with the default value: "{}". The code inside
`Spree::Preferences::Preferable` that handles type conversion throws an
error if the input for a preference with type :hash is not a Hash
object.

This change allows for a well-formed hash to be input as a string and
converts it to a Hash before saving the record.