Handle processor and gateway error responses
The existing code here will only return error messages if they come from
Braintree. In the case that the transaction is rejected by the gateway
or the processor, we'll return an error response with a blank message.

For these cases, the `errors` array will be empty, and we need to look
at the gateway rejection status or processor response codes as per the
documentation here:
https://developers.braintreepayments.com/reference/response/transaction/ruby#result-object