Replace the routes helper
This was added in 4d39dba85a461508b376456b9d2fbe672ca371ce

The relevant part of the commit message is:

```
    One issue encountered is that Solidus doesn't always use the `spree.`
    routing proxy explicitly. If a page is rendered from a controller inside
    the gem, any url helpers that don't explicitly use a routing proxy will
    be looked up in the gem's routes causing errors. The RoutesHelper module
    was added to deal with that. When included in a controller, the
    method_missing will retry the url helper using the spree routing proxy
    and return that result if it exists, otherwise it will raise the error.
```

We shouldn't be obfuscating this from users. If users want to use engine
routes, they should be using the relevant engine route helper, as
provided by isolate namspace and documented in
http://guides.rubyonrails.org/engines.html#routes

This problem arises from the fact the admin layout uses implicit mounted
routes in older versions of Solidus. To get around this, we add sprees
routes as a helper. This is how Rails' mounted routeset work.
