Always use the same zipcode in address factory
The Solidus address factory randomizes the zipcode. The OrderWalkThrough we use in the credit card checkout spec uses this factory for the user addresses.

For credit card payments we transmit the billing address to braintree, for paypal payments the shipping address.

As we match the body in our VCR settings VCR can not match the request anymore and therefore cannot replay existing cassettes.

By modifying the spree address factory we ensure that the zipcode is always the same.