Validate payment_type on source
The payment_type is used to differ payment soruces. I.E. lots of query methods uses this column to decide which kind of payment this source instance is, even presentor methods uss this clumn to look up the human name in the translation files by this column.

Therefore it is necessary to validate its presence and add a not null constrained in the database.