Change empty form to invalid form
Submitting an empty form as an exemplar of invalid input was causing some timing issues, with Capybara submitting the form before the Braintree iframes were set up correctly.

Entering an invalid card number fulfills the point of the spec – confirming that everything is reset correctly after an invalide entry – but ensures that Capybara has to wait until the form is actually ready to submit.

A more complete solution to this problem would be to keep the submit button disabled until after the iframes are in place, but I’ll raise that as a separate issue.