Remove integer type declaration from cassettes
Lots of older cassettes had `type="integer"` attributes in the xml
of the request body. Newer cassettes do not have this. This causes
build errors depending on which machine the specs run.

It is not 100% clear what's the cause of this.

Maybe the builder gem? - braintree depends on >=2.0.0 while 3.2.3 is the most recent version
Maybe the braintree gem? - 2.76.0 is the most recent version and was updated lately
Maybe libxml? - The version on the different machines may vary