Add spec for omitting the CVV response code
For PCI compliance reasons we do not send the CVV with the transaction. The hosted fields frontned code provided by Braintree already checks the CVV for us.

That's why we always get "I" (for "Not Provided") back. Solidus considers this as risky and marks the order as such. This will confuse admin users and therefore we omit this value.