Implement PaymentMethod#try_void
Solidus >= 2.4 introduced a new way to cancel payments. Instead of always
expecting the payment method to void the payment it now asks to try a
void and create a refund if that is not possible.

We always made this differentiation but returned a void response in both cases,
because this is what Solidus was expecting before v2.4.0.

We're now embracing the new API by implementing `try_void`.