Rescue Braintree void errors and refund instead
Sometimes Braintree returns a voidable transaction status although it is not voidable anymore.

According to them there is a hidden state called `settling` that is not returned by the API for a transaction status request.

When we then void we get back an error from the API that leaves the transaction in an unvoided and not-refunded state.

This is solved by catching the error and let Solidus create a refund instead.