Restart checkout when cart Paypal button is used
Every time an express Paypal checkout is initiated from the cart we must reset
the checkout flow.
This is needed because the checkout may be in any state and the newly started
Paypal checkout may send us a new, different address, thus we must recalculate
all the things (taxes, shipping methods etc).