Persist credit card type and last digits on sources
We were always getting this information back in the payments responses,
but did not have fields to store it in on the
`SolidusPaypalBraintree::Source` objects. In a previous commit I added a
`cc_type` and `last_digits` field to sources to mimic
`Spree::CreditCard`s so that we can store this information.

There is no reason this information should ever change after the initial
request since any new card information would result in a new source.
That means we can store it right away and avoid making an API call every
time we need to look up the card type or last 4 digits.