Stop importing address from PayPal
Closes #226. When PayPal sends the authorization response to the server,
it format the address with a single field for the first name and the
last name, `recipientName`. We can’t safely infer the first name and the
last name from that field (see #226), so we have to skip the user
address updating. The good news is that the user address updating in
this case is redundant, since Solidus always requires the address to be
set before the payment happens, and such address is sent to PayPal with
the `shippingAddressEditable` configuration set to `false`, so the
address is necessarily the same as the one stored anyway.