Fix 3DS testing
For whatever reason, the listed Braintree cards that are _supposed_
to trigger the fail/pass conditions do not anymore. Instead, using the
cards listed in Cardinal documentation does trigger those conditions.

The structure of the 3ds window is also slightly different, with the
input being nested in two iframes now.