Update to New PayPal SDK
This switches from the braintree SDK to the new PayPal SDK, which
allows for displaying credit messaging/buttons.

This update adds two new config options:
* Layout, accepting either `horizontal` or `vertical`, changing
the layout of the buttons
* Messaging, a boolean, that controls whether the messaging
component is loaded in.
This update also removes the `size` config option, as it no longer
does anything. PayPal uses the responsive size for everything now,
and expects you to resize your elements locally to whatever size
you want.