Move decorators to the correct folders and add EngineExtensions
The standard way of handling decorators is to move them into a folder
for what they decorate - in this case, `models`, and using
EngineExtensions in `engine.rb` which will grab your decorators from the
folder directly under `decorators` - meaning that you don't need to
include `models` in the namespace of the file.

This also moves the decorators located under `controllers` to the
decorators folder.