Use messaging partial on payment buttons
Uses the existing messaging partial on the cart & checkout payment
buttons, and passes the total & placement along. Also adds a
`skip_initialization` option to the partial, because we don't want to
initialize the messaging component twice, and PayPal _really_ doesn't
like loading in their SDK twice, even if for two different components.