Add force_buyer_country payment method preference for testing
Resolves issue/ticket:
- https://github.com/solidusio/solidus_paypal_braintree/issues/303

Usually on checkout, PayPal will look at the buyer's location by their
ip and determine which funding sources are available to them. For
example, allow Venmo as an option for US buyers, but not for others.

With this new preference, you can set a country to override the buyer
location. Allowing you on non-production environments to see the
different funding sources for different customers and its integration.

More information:
https://developer.paypal.com/docs/business/javascript-sdk/javascript-sdk-configuration/#buyer-country