Add email to Braintree customer
Issue: https://github.com/solidusio/solidus_paypal_braintree/issues/309

Braintree comes with free basic fraud protection. One of the tools
is Risk Threshold Rules (velocity checks) [1]. One of the checks is
checking amount of transactions per customer email. However, the email
was not being attached to the customer when creating it before the
transaction.

This fixes that so developers can utilize this check.

[1] https://developer.paypal.com/braintree/articles/guides/fraud-tools/basic/risk-threshold-rules