Add paypal_funding_source to Source model
Saving the data which fundingSource, i.e. which financing option was
used for the PayPal payment. This data be shown on the frontend to
confirm to the user how they paid, for example, Venmo.

This is reassurance and was also a requirement set by PayPal for our
integration of Venmo.

Note:
- there is no presence validator it is only useful for PayPal sources
  and not Credit Card and ApplePay sources.
- a save callback has been made to ensure it is nil for Credit Card and
  ApplePay sources. Otherwise, a user could click on a PayPal button
  (which sets the fundingSource), and then decide to use the hosted form
  for a credit card. This would then set the fundingSource to 'paypal'
  if this callback didn't exist.