Do not safeguard including of Spree::Preferences::Persistable
The safeguard prevents this module to be included because the
class responds to preference (it is included in Spree::Base).

My guess is that this safeguard was included to avoid including
Spree::Preferences::Preferable twice (it is included by Spree::Preferences::Persistable)

But Ruby will take care of not including the same module twice.