Fix broken specs due to Solidus factory changes
Due to the following PR, the Country factory no longer always returns
true for `states_required`. It can be true for the appropriate countries
that require a state in shipping addresses. Because of this, addresses
using a country where `states_required` is true, such as the US, need
a state in the Spree::Address record to be valid.
https://github.com/solidusio/solidus/pull/4272