Fix 3D Secure specs
Cardinal, that handles "3D Secure" for Braintree made some changes to
how the legacy 3Ds test numbers are handled. Because of this, the
numbers were changed to trigger the correct scenarios needed. The
DOM of the Cardinal iFrame also changed and needed tweaking. Cardinal
also likes the expiration date to be January and 3 years ahead of
today's year.

References:
- https://cardinaldocs.atlassian.net/wiki/spaces/CCen/pages/400654355/3DS+1.0+Test+Cases
- https://cardinaldocs.atlassian.net/wiki/spaces/CCen/pages/903577725/EMV+3DS+Test+Cases#Test-Case-12%3A-Error-on-Authentication