Fix bin/sandbox script generator
Because this extension decorates the User class, the original User class
must be loaded before the decorator, otherwise the sandbox migrations
fail. To fix this, solidus_auth_devise will be listed in the Gemfile
before extension.

The `--with-authentication` flag will still be given to the Solidus
generator so that includes solidus_auth_devise' seeds, however, Solidus
will not install it twice.
https://github.com/solidusio/solidus/blob/efe2b3a20d0577413e4810c008ea564e637c36a3/core/lib/generators/solidus/install/app_templates/authentication/devise.rb