Install solidus_braintree using the install generator
Mainly this is done to let specs pass. In fact this command is going
to update the spec/dummy app to run the braintree generator and have
all the file needed in place.

But it's fine to have this available for everyone. If someone wrongly
install the extension or run the install generator after upgrading
SolidusPaypalBraintree to 2.0, this installer will actually make it
work.